refactor: adjust by code review
